test(testermint): exclude classic-inference tests removed by dapi deprecation
Tag tests that depend on the removed classic inference flow with
@Tag("exclude") so listAllTestClasses drops their suites from the CI
matrix entirely:

- Whole class excluded (matrix job skipped): ConsumerTests,
  InferenceAccountingTests, InferenceFailureAccountingTests,
  ValidationTests, StreamingInferenceTests, DynamicPricingTest,
  BandwidthLimiterTests, StatsApiE2ETest, InferenceRetryTests,
  MultiModelPoCTests, NodeDisableInferenceTests
- Per-test exclusion in classes that still have passing tests:
  InferenceTests (11 of 18), MultiModelTests, PruningTests
- InvalidationTests: all cluster tests excluded; its pure unit test
  (logprob hash) moved to PromptHashingTests, so the suite also drops
  out of the matrix

Suites that only need their inference glue rewritten (Governance,
Restrictions, TransactionFee, NodeAdminState, Collateral, Participant,
Upgrade) are intentionally untouched and will be fixed separately.
