+++ testermint/src/test/kotlin/BandwidthLimiterTests.kt
@@ -2,11 +2,14 @@ import com.github.kittinunf.fuel.core.FuelError
 import com.productscience.*
 import com.productscience.data.*
 import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.Tag
 import org.junit.jupiter.api.Test
 import org.tinylog.kotlin.Logger
 import java.time.Instant
 import kotlin.test.assertNotNull
 
+// Classic inference flow was removed (PR #1386); these tests exercise the deprecated endpoints.
+@Tag("exclude")
 class BandwidthLimiterTests : TestermintTest() {
 
     @Test
